main - move pipeline to cicd module
